omie-tenant-manager/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
bcrypt>=4.1.1
pyjwt>=2.8.0
python-multipart>=0.0.9
schedule>=1.2.0